Riscritta la funzione train.c per creare una lista concatenata, presenta una funzione per liberare la memoria allocata, ad essa è necessario passare la struct immagine *testa, cioè un puntatore alla prima struct della lista, altrimenti non si avrà una completa liberazione della memoria, minori modifiche a print e pixel_at di image_helper.c, modificata la funzione image_char2double di mnist.c

+++ lib/image_helper/include/image_helper.h
@@ -9,7 +9,7 @@ struct immagine {
 	struct immagine *next;
 };
 
-int pixel_at(int x, int y, const struct immagine *image);
+double pixel_at(int x, int y, const struct immagine *image);
 
 void set_label(int x, struct immagine *image);
